Former-commit-id: b0c97a8a9a85bd93b9983f2a9970b9ed4f4d52b7

+++ back/src/main/java/site/chachacha/fitme/domain/product/entity/Product.java
@@ -118,7 +118,7 @@ public class Product extends BaseEntity {
         this.likeCount = this.likeCount + 1;
     }
 
-    public void updateMonthlyPopularityScore(int monthlyPopularityScore) {
+    public void updateMonthlyPopularityScore(double monthlyPopularityScore) {
         this.monthlyPopularityScore = monthlyPopularityScore;
     }
